Rename `CreatedByYou` component to be more accurate
